DAC working and lots of new files created!

+++ source/node2/pwm_driver.c
@@ -1,4 +1,8 @@
 #include "pwm_driver.h"
+#include "sam.h"
+#include "timer.h"
+
+#include <stdint.h>
 
 #define RC_20MS 840000
 #define RA_MID  63000
@@ -14,8 +18,8 @@ void pwm_init(void)
     timer_pwm_init();
 }
 
-void pwm_set_duty_cycle(float percent) {
-    uint32_t duty_cycle = DUTY_CYCLE_INTERCEPT + DUTY_CYCLE_SLPE*percent;
+void pwm_set_duty_cycle(uint8_t percent) {
+    uint32_t duty_cycle = DUTY_CYCLE_INTERCEPT + DUTY_CYCLE_SLPE * percent;
     if (duty_cycle < RA_MAX || duty_cycle > RA_MIN) {
         TC0->TC_CHANNEL[0].TC_RA = duty_cycle;
     }
